+++ src/main/kotlin/racingcar/car/engine/ConstantEngine.kt
@@ -1,4 +1,4 @@
-package racingcar.engine
+package racingcar.car.engine
 
 class ConstantEngine(private val advancePoint: Int) : Engine {
     override fun advancePoint(): Int {
