+++ src/main/kotlin/racingcar/car/engine/Engine.kt
@@ -1,4 +1,4 @@
-package racingcar.engine
+package racingcar.car.engine
 
 interface Engine {
     fun advancePoint(): Int
