+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -3,6 +3,10 @@ package racingcar.view
 import racingcar.count.CarCount
 import racingcar.count.MoveCount
 
+private fun printBy(askCount: Int) {
+    if (askCount % 4 == 0) println("같은걸 $askCount 번째 물어봅니다... s(^_^)a")
+}
+
 class InputView(private val minCarCount: Int, private val minMoveCount: Int) {
     val carCount = getCarCount()
     val moveCount = getMoveCount()
@@ -23,11 +27,5 @@ class InputView(private val minCarCount: Int, private val minMoveCount: Int) {
     } catch (e: IllegalArgumentException) {
         println(e.message)
         getMoveCount(askCount + 1)
-    }
-
-    companion object {
-        fun printBy(askCount: Int) {
-            if (askCount % 4 == 0) println("같은걸 $askCount 번째 물어봅니다... s(^_^)a")
-        }
     }
 }
